requirements.txt
blinker==1.9.0
click==8.1.8
flask==3.0.3
Flask-Login==0.6.3
importlib-metadata==8.7.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
pymysql==1.1.2
python-dotenv==1.0.1
werkzeug==3.0.6
zipp==3.23.0
